Add AnalogKillSwitch class, a subclass of AnalogInput
AnalogKillSwitch works like AnalogInput, but also sends out a current
on Pin 18. This goes through the kill switch relay in the front of the
sub, and when the sub is alive, the circuit is completed and we can
read it on Pin 16. So it works like a continuity tester. Also,
AnalogKillSwitch does some basic averaging on the data because AnalogIn
occasionally returns noisy data.

Also change AnalogInput's private things to protected, so that
AnalogInput's subclasses (such as AnalogKillSwitch) can access those
things.